fix(ci): host を build してデモ build を通し、Android viewport 型不整合を解消
2件の既存 CI 失敗を修正する。

1) Deploy demo: deploy-pages.yml に `build:host` を追加。#477 で host
   bootstrap を @hayate/host(exports は dist のみ)へ分離した際にワークフローへ
   build:host を入れ忘れ、CI で dist が生成されず Rolldown が
   `import('@hayate/host')` を解決できず todo build が落ちていた。

2) Android cross-compile: app_tsubame.rs の last_viewport を
   Option<(u32,u32)> → Option<(f32,f32)> に。viewport_for_surface が
   (f32,f32) を返し set_viewport も f32 を取るよう f32 化されたのに、
   last_viewport の型注釈だけ u32 のままで E0308 になっていた。

+++ Hayate/crates/platform/mobile/android/src/app_tsubame.rs
@@ -118,7 +118,7 @@ pub(crate) fn run(app: AndroidApp) {
     };
 
     // 直近の viewport。reload で tree を作り直したら、新 tree に再適用して描画サイズを引き継ぐ。
-    let mut last_viewport: Option<(u32, u32)> = None;
+    let mut last_viewport: Option<(f32, f32)> = None;
 
     // ── full reload 購読（#533）────────────────────────────────────────────────
     // WS `reload` 受信で reload フラグを立て、poll ループが拾って再 boot する。WS の blocking read は
